+++ src/ska_tmc_dishleafnode/commands/abort_command.py
@@ -104,6 +104,8 @@ class Abort(DishLNCommand):
             ]
             self.command_uniq_id = ""
         self.component_manager.clear_configure_command_events_flags()
+        self.component_manager.receiver_band = ""
+        self.component_manager.update_rxband_health_aggregation()
 
     # pylint: disable=arguments-differ