+++ src/ska_tmc_dishleafnode/commands/trackstop_command.py
@@ -62,6 +62,8 @@ class TrackStop(DishLNCommand):
                 self.command_uniq_id
             ]
             self.command_uniq_id = ""
+        self.component_manager.receiver_band = ""
+        self.component_manager.update_rxband_health_aggregation()
 
     # pylint: disable=unused-argument
     @timeout_tracker