+++ tests/integration/test_apply_pointing_model.py
@@ -12,8 +12,11 @@ from ska_tmc_common.dev_factory import DevFactory
 from tests.settings import (
     DISH_LEAF_NODE_DEVICE,
     DISH_MASTER_DEVICE,
+    DISHLN_POINTING_DEVICE,
+    log_and_assert_health,
     logger,
     wait_and_validate_attribute_value_available,
+    wait_for_attribute_health_value,
 )
 
 
@@ -26,7 +29,7 @@ def gpm_validation_result(
     while timeout >= 0:
         attribute_value = group_callback[
             "gpmValidationResult"
-        ].assert_change_event(Anything, lookahead=1,)["attribute_value"]
+        ].assert_change_event(Anything, lookahead=5,)["attribute_value"]
         gpm_validation_result = json.loads(attribute_value)
         if gpm_validation_result[band_name] == band_result:
             flag = True
@@ -43,6 +46,7 @@ def apply_pointing_model(tango_context, dishln_name, group_callback, gpm_json):
     dev_factory = DevFactory()
     dish_leaf_node = dev_factory.get_device(dishln_name)
     dish_master_dev = dev_factory.get_device(DISH_MASTER_DEVICE)
+    dishln_pointing_device = dev_factory.get_device(DISHLN_POINTING_DEVICE)
     dish_leaf_node.subscribe_event(
         "longRunningCommandResult",
         tango.EventType.CHANGE_EVENT,
@@ -96,15 +100,15 @@ def apply_pointing_model(tango_context, dishln_name, group_callback, gpm_json):
 
     group_callback["longRunningCommandResult"].assert_change_event(
         (unique_id[0], '[0, "Successfully wrote the GPM values"]'),
-        lookahead=8,
+        lookahead=5,
     )
     group_callback["globalPointingModelParams"].assert_change_event(
         Anything,
-        lookahead=2,
+        lookahead=5,
     )
     group_callback["gpmVersion"].assert_change_event(
         Anything,
-        lookahead=2,
+        lookahead=5,
     )
     gpm_version = json.loads(dish_leaf_node.gpmversion)
     assert gpm_version['Band_1'] == 'main'
@@ -116,7 +120,7 @@ def apply_pointing_model(tango_context, dishln_name, group_callback, gpm_json):
 
     group_callback["healthState"].assert_change_event(
         HealthState.OK,
-        lookahead=2,
+        lookahead=5,
     )
 
     # Scenario 2:
@@ -130,14 +134,22 @@ def apply_pointing_model(tango_context, dishln_name, group_callback, gpm_json):
     dish_master_dev.band1pointingmodelparams = dish_band1pointingmodelparams
     group_callback["globalPointingModelParams"].assert_change_event(
         Anything,
-        lookahead=2,
+        lookahead=5,
     )
 
     gpm_validation_result(group_callback, "Band_1", "FAILED")
 
     group_callback["healthState"].assert_change_event(
         HealthState.DEGRADED,
-        lookahead=2,
+        lookahead=5,
+    )
+
+    log_and_assert_health(
+        dish_leaf_node,
+        dish_master_dev,
+        dishln_pointing_device,
+        HealthState.DEGRADED,
+        "GPM validation failed for GPM index 0.",
     )
     gpm_version = json.loads(dish_leaf_node.gpmversion)
     assert gpm_version['Band_1'] == 'main'
@@ -149,12 +161,12 @@ def apply_pointing_model(tango_context, dishln_name, group_callback, gpm_json):
     dish_master_dev.band1pointingmodelparams = dish_band1pointingmodelparams
     group_callback["globalPointingModelParams"].assert_change_event(
         Anything,
-        lookahead=2,
+        lookahead=5,
     )
     gpm_validation_result(group_callback, "Band_1", "OK")
     group_callback["healthState"].assert_change_event(
         HealthState.OK,
-        lookahead=2,
+        lookahead=5,
     )
     gpm_version = json.loads(dish_leaf_node.gpmversion)
     assert gpm_version['Band_1'] == 'main'
@@ -165,14 +177,22 @@ def apply_pointing_model(tango_context, dishln_name, group_callback, gpm_json):
     dish_master_dev.band3pointingmodelparams = dish_band1pointingmodelparams
     group_callback["globalPointingModelParams"].assert_change_event(
         Anything,
-        lookahead=2,
+        lookahead=5,
     )
 
     gpm_validation_result(group_callback, "Band_3", "FAILED")
 
     group_callback["healthState"].assert_change_event(
         HealthState.DEGRADED,
-        lookahead=2,
+        lookahead=5,
+    )
+
+    log_and_assert_health(
+        dish_leaf_node,
+        dish_master_dev,
+        dishln_pointing_device,
+        HealthState.DEGRADED,
+        "GPM validation failed for GPM index 2.",
     )
     gpm_version = json.loads(dish_leaf_node.gpmversion)
     assert gpm_version['Band_3'] == 'UNKNOWN'
@@ -195,19 +215,17 @@ def apply_pointing_model(tango_context, dishln_name, group_callback, gpm_json):
     dish_master_dev.band5apointingmodelparams = np.array([])
     group_callback["globalPointingModelParams"].assert_change_event(
         Anything,
-        lookahead=2,
+        lookahead=5,
     )
     group_callback["gpmVersion"].assert_change_event(
         Anything,
-        lookahead=2,
+        lookahead=5,
     )
 
     gpm_validation_result(group_callback, "Band_5a", "OK")
 
-    group_callback["healthState"].assert_change_event(
-        HealthState.DEGRADED,
-        lookahead=2,
-    )
+    wait_for_attribute_health_value(dish_leaf_node, "healthState", 1)
+
     # Assert Band_5a contains values
     assert (dish_master_dev.band5apointingmodelparams).tolist()
 
@@ -235,7 +253,7 @@ def apply_pointing_model(tango_context, dishln_name, group_callback, gpm_json):
         "longRunningCommandResult"
     ].assert_change_event(
         (unique_id[0], Anything),
-        lookahead=8,
+        lookahead=5,
     )[
         "attribute_value"
     ]
@@ -252,7 +270,7 @@ def apply_pointing_model(tango_context, dishln_name, group_callback, gpm_json):
 
     group_callback["healthState"].assert_change_event(
         HealthState.OK,
-        lookahead=2,
+        lookahead=5,
     )
 
 
@@ -279,7 +297,7 @@ def ApplyPointingModel_with_invalid_tm_path(
         "longRunningCommandResult"
     ].assert_change_event(
         (unique_id[0], Anything),
-        lookahead=4,
+        lookahead=5,
     )[
         "attribute_value"
     ]
@@ -311,7 +329,7 @@ def ApplyPointingModel_with_invalid_dish_id(
         "longRunningCommandResult"
     ].assert_change_event(
         (unique_id[0], Anything),
-        lookahead=4,
+        lookahead=5,
     )[
         "attribute_value"
     ]
@@ -343,7 +361,7 @@ def ApplyPointingModel_with_invalid_json(
         "longRunningCommandResult"
     ].assert_change_event(
         (unique_id[0], Anything),
-        lookahead=4,
+        lookahead=5,
     )[
         "attribute_value"
     ]
@@ -389,11 +407,11 @@ def gpm_version_restart_scenario(
 
         group_callback["longRunningCommandResult"].assert_change_event(
             (unique_id[0], '[0, "Successfully wrote the GPM values"]'),
-            lookahead=8,
+            lookahead=5,
         )
         group_callback["gpmVersion"].assert_change_event(
             Anything,
-            lookahead=2,
+            lookahead=5,
         )
 
     dish_leaf_node.init()
@@ -406,7 +424,7 @@ def gpm_version_restart_scenario(
 
     assertion_data = group_callback["gpmVersion"].assert_change_event(
         Anything,
-        lookahead=2,
+        lookahead=5,
     )
 
     gpm_version_data = json.loads(assertion_data["attribute_value"])