+++ CHANGELOG.md
@@ -6,11 +6,21 @@ All notable changes to this project will be documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 
+[0.29.2]
+********
+Added
+-----
+* Implemented  HealthInfo on Dish Leaf Node
+* Improved visibility: failures (e.g. GPM/KValue validation, Dish Manager state, band capability) now surface in healthInfo .
+
+
 [0.29.1]
 ********
 Added
 -----
+
 * Added fix for GPM issue faced on mid integration.
+
 
 [0.29.0]
 ********